Refactoring: move postings reading utilities to dedicated file
