+++ pkg/storegateway/bucket_test.go
@@ -11,7 +11,6 @@ package storegateway
 import (
 	"bytes"
 	"context"
-	"encoding/binary"
 	"fmt"
 	"math"
 	"math/rand"
@@ -1571,24 +1570,6 @@ func mustMarshalAny(pb proto.Message) *types.Any {
 		panic(err)
 	}
 	return out
-}
-
-func TestBigEndianPostingsCount(t *testing.T) {
-	const count = 1000
-	raw := make([]byte, count*4)
-
-	for ix := 0; ix < count; ix++ {
-		binary.BigEndian.PutUint32(raw[4*ix:], rand.Uint32())
-	}
-
-	p := newBigEndianPostings(raw)
-	assert.Equal(t, count, p.length())
-
-	c := 0
-	for p.Next() {
-		c++
-	}
-	assert.Equal(t, count, c)
 }
 
 func createBlockWithOneSeriesWithStep(t test.TB, dir string, lbls labels.Labels, blockIndex, totalSamples int, random *rand.Rand, step int64) ulid.ULID {
